Refactor: share unified deposit proof helper across proposer backends
