+++ nightfall_proposer/src/driven/mock_prover.rs
@@ -5,18 +5,13 @@ use itertools::{izip, Itertools};
 use jf_plonk::{
     errors::PlonkError,
     nightfall::ipa_structs::VerifyingKey,
-    nightfall::FFTPlonk,
-    proof_system::{RecursiveOutput, UniversalRecursiveSNARK},
+    proof_system::RecursiveOutput,
     recursion::circuits::Kzg,
-    transcript::RescueTranscript,
 };
-use jf_primitives::{pcs::prelude::UnivariateKzgPCS, rescue::sponge::RescueCRHF};
 use jf_utils::fr_to_fq;
 use lib::{
-    error::UnifiedProofError,
     merkle_trees::trees::{MerkleTreeError, MutableTree, TreeMetadata},
-    nf_client_proof::{PrivateInputs, PublicInputs},
-    plonk_prover::circuits::unified_circuit::unified_circuit_builder,
+    nf_client_proof::PublicInputs,
     plonk_prover::{get_client_proving_key, plonk_proof::PlonkProof},
     shared_entities::DepositData,
 };
@@ -30,39 +25,14 @@ use mongodb::{bson::doc, Client};
 use super::rollup_prover::RollupProofError;
 use crate::{
     domain::entities::ClientTransactionWithMetaData,
-    driven::rollup_prover::Bn254Output,
+    driven::{rollup_prover::Bn254Output, unified_deposit_prover::create_unified_deposit_proof},
     initialisation::get_db_connection,
     ports::proving::RecursiveProvingEngine,
     ports::trees::{CommitmentTree, HistoricRootTree, NullifierTree},
 };
-use ark_bn254::{Bn254, Fq as Fq254, Fr as Fr254};
+use ark_bn254::{Fq as Fq254, Fr as Fr254};
 use ark_std::Zero;
 pub struct MockProver;
-
-impl MockProver {
-    #[allow(dead_code)]
-    fn create_unified_deposit_proof(
-        deposit_data: &[DepositData; 4],
-        public_inputs: &mut PublicInputs,
-    ) -> Result<PlonkProof, UnifiedProofError> {
-        let mut private_inputs = PrivateInputs::for_deposit(deposit_data);
-        *public_inputs = PublicInputs::for_deposit();
-        let mut circuit = unified_circuit_builder(public_inputs, &mut private_inputs)
-            .map_err(UnifiedProofError::from)?;
-        circuit
-            .finalize_for_recursive_arithmetization::<RescueCRHF<Fq254>>()
-            .map_err(UnifiedProofError::from)?;
-        let pk = get_client_proving_key();
-
-        let output = FFTPlonk::<UnivariateKzgPCS<Bn254>>::recursive_prove::<
-            _,
-            _,
-            RescueTranscript<Fr254>,
-        >(&mut ark_std::rand::thread_rng(), &circuit, pk, None, true)
-        .map_err(UnifiedProofError::from)?;
-        Ok(PlonkProof::from_recursive_output(output, &pk.vk))
-    }
-}
 
 impl RecursiveProvingEngine<PlonkProof> for MockProver {
     type PreppedInfo = Fr254;
@@ -280,6 +250,6 @@ impl RecursiveProvingEngine<PlonkProof> for MockProver {
         deposit_data: &[DepositData; 4],
         public_inputs: &mut PublicInputs,
     ) -> Result<PlonkProof, Self::Error> {
-        Self::create_unified_deposit_proof(deposit_data, public_inputs).map_err(Self::Error::from)
+        create_unified_deposit_proof(deposit_data, public_inputs).map_err(Self::Error::from)
     }
 }
